+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - Routing patterns can be ignored, and grouped together when reporting to Prometheus.
+- Endpoint label behavior can be altered with the new `EndpointLabel` enum.
 - Added a new builder `PrometheusMetricLayerBuilder` to easily customize these.
 
   ```rust
@@ -15,11 +16,14 @@ All notable changes to this project will be documented in this file.
       .with_ignore_patterns(&["/foo", "/sensitive"])
       // if the any of the second argument matches, report them at the `/bar` endpoint
       .with_group_patterns_as("/bar", &["/foo/:bar", "/foo/:bar/:baz"])
+      // use `axum::extract::MatchedPath`, and if that fails, use the exact requested URI
+      .with_endpoint_label_type(EndpointLabel::MatchedPath)
       .with_default_metrics()
       .build_pair();
   ```
 
-- A [builder-example](examples/builder-example/).
+- A [builder-example](examples/builder-example/) and an [endpoint-type-example](examples/endpoint-type-example/).
+
 - The metric names can be changed by setting some environmental variables at compile time. It is best to set these in the `config.toml` (note this is not the same file as `Cargo.toml`):
   ```toml
   [env]